requirements.txt
# Project Dependencies
# This file lists all dependencies for the Embedded Finance Platform project.

# ==========================================
# BACKEND (Python)
# Location: backend/backend_requirements.txt
# ==========================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Background Tasks
celery==5.3.4
redis==5.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# ML & Analytics
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Monitoring & Logging
structlog==23.2.0
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Environment
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Validation
email-validator==2.1.0

# Rate Limiting
slowapi==0.1.9

# Caching
aioredis==2.0.1


# ==========================================
# AI AGENT (Python)
# Location: ai_agent/ai_requirements.txt
# ==========================================

# Core HTTP and async support
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Data handling and validation
pydantic==2.5.2
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# JSON and data processing
ujson==5.9.0
orjson==3.9.10

# Retry and error handling
tenacity==8.2.3
backoff==2.2.1

# Optional: Database integration
asyncpg==0.29.0
redis==5.0.1
aioredis==2.0.1

# Optional: Metrics and monitoring
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8
passlib==1.7.4


# ==========================================
# FRONTEND (Node.js/React)
# Location: frontend/package.json
# ==========================================

# Dependencies
@heroicons/react: ^2.2.0
@reduxjs/toolkit: ^2.8.2
@testing-library/jest-dom: ^5.17.0
@testing-library/react: ^13.4.0
@testing-library/user-event: ^13.5.0
@types/jest: ^27.5.2
@types/react: ^18.3.0
@types/react-dom: ^18.3.0
apexcharts: ^4.7.0
axios: ^1.10.0
dompurify: ^3.2.6
lucide-react: ^0.511.0
marked: ^15.0.12
react: ^18.3.1
react-apexcharts: ^1.4.1
react-dom: ^18.3.1
react-hot-toast: ^2.5.2
react-redux: ^9.2.0
react-router-dom: ^6.30.1
react-scripts: 5.0.1
recharts: ^2.15.4
styled-components: ^6.1.9
typescript: ^4.9.5
web-vitals: ^2.1.4

# DevDependencies
@tailwindcss/forms: ^0.5.10
@tailwindcss/typography: ^0.5.19
@types/node: ^20.17.56
@types/styled-components: ^5.1.34
autoprefixer: ^10.4.22
postcss: ^8.5.6
tailwindcss: ^3.4.17
vite: ^6.3.5
